backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
requests
beautifulsoup4
aiohttp
httpx
python-multipart
python-dateutil
tenacity
scikit-learn
numpy
pytest
pytest-asyncio
